pyproject.toml
[project]
name = "ai-doc-gen"
version = "1.0.0"
description = "AI-powered code documentation generator that analyzes repositories and creates comprehensive documentation"
authors = [{ name = "Milad Noroozi", email = "norooziosos@gmail.com" }]
requires-python = ">=3.13,<3.14"
dependencies = [
    "pydantic>=2.11.7",
    "psutil>=7.0.0",
    "jinja2>=3.1.5",
    "ujson>=5.10.0",
    "pydantic-ai[logfire]>=0.4.2",
    "nest-asyncio>=1.6.0",
    "python-gitlab>=6.1.0",
    "gitpython>=3.1.44",
    "logfire>=3.24.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=9.4.0",
]

[project.scripts]
ai-doc-gen = "src.main:cli_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.sources]
"src" = "src"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"
